refactor: extract types and utilities from minimal serializer ES module
