+++ packages/router-component-store/src/lib/minimal-route-data.ts
@@ -1,0 +1,9 @@
+import { Data } from '@angular/router';
+import { OmitSymbolIndex } from './util-types/omit-symbol-index';
+
+/**
+ * Serializable route `Data` without its symbol index, in particular without the
+ * `Symbol.for(RouteTitle)` key as this is an internal value for the Angular
+ * `Router`.
+ */
+export type MinimalRouteData = OmitSymbolIndex<Data>;
