+++ packages/router-component-store/src/lib/polyfills/object.ts
@@ -1,0 +1,11 @@
+/**
+ * `Object.fromEntries` polyfill.
+ */
+export function objectFromEntries<TValue>(entries: [string, TValue][]): {
+  [key: string]: TValue;
+} {
+  return entries.reduce(
+    (object, [key, value]) => ({ ...object, [key]: value }),
+    {}
+  );
+}
